% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_formula.R
\name{create_formula}
\alias{create_formula}
\title{Create a formula with many covariates for the HABs drivers project using spmodel}
\usage{
create_formula(dep_var, ind_vars, log = FALSE, const = 0, interact = NULL)
}
\arguments{
\item{dep_var}{The dependent variable of the model as a string.}

\item{ind_vars}{A vector string of covariate names.}

\item{log}{TRUE/FASLE Should the dependent variable be log10 transformed. Default is FALSE.}

\item{const}{Numeric. If log-transforming, a constant to add to the dependent variable. Default is 0.}

\item{interact}{Any interactions to include in the model taking the string form of "var_name1 * var_name2". Default is NULL.}
}
\value{
A formula object
}
\description{
Create a formula with many covariates for the HABs drivers project using spmodel
}
\examples{
create_formula(
  dep_var = "B_G_DENS",
  ind_vars = c("NTL", "MAXDEPTH", "agr_ws"),
  log = TRUE,
  const = 1000
 )
}
