% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{eco_ag3_simpl}
\alias{eco_ag3_simpl}
\title{Aggregated ecoregions for the conterminous US}
\format{
A sf object with 3 rows and 3 variables:
\describe{
\item{WSA_3}{The ecoregion code (character)}
\item{WSA_3_NM}{The ecoregion name (character)}
\item{geometry}{Shape information (multipolygon)}
}
}
\source{
{HABsDrivers} Package for estimating lake HABs risk.
}
\usage{
eco_ag3_simpl
}
\description{
A shapefile containing the three aggregated ecoregions for the conterminous US.
Note this shapefile has been simplified to a reduced number of vertices and is
only intended for data visualization purposes at the national scale.
}
\keyword{datasets}
