% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HABsDrivers-package.R
\docType{package}
\name{HABsDrivers-package}
\alias{HABsDrivers}
\alias{HABsDrivers-package}
\title{HABsDrivers: Patterns and Predicting CyanoHABs in US Lakes}
\description{
Analysis of the factors associated with harmful algal bloom (HAB) risk. This project draws from EPA dataset to assess the risk of HABs in lakes across the contiguous US.
}
\author{
\strong{Maintainer}: Amalia Handler \email{handler.amalia@epa.gov} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
