% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{PredData}
\alias{PredData}
\title{Compiled data for the lake HABs risk modeling project}
\format{
A data frame with 124529 rows and 27 variables:
\describe{
\item{COMID}{A site ID assigned to a lake (character)}
\item{DSGN_CYCLE}{The year of the NLA survey cycle (2007, 2012, 2017)}
\item{UNIQUE_ID}{A unique ID assigned to each lake sampled across NLAs (NLA_STATE-NUMBER)}
\item{n_dev_inputs}{Annual total nitrogen inputs from development (non-farm) into the watershed, mean of 2002, 2007,
and 2012 National Nutrient Inventory (kg N/ha)}
\item{n_farm_inputs}{Annual total nitrogen inputs from agricultural activities into the watershed, mean of 2002, 2007,
and 2012 National Nutrient Inventory (kg N/ha)}
\item{p_dev_inpts}{Annual total phosphorus inputs from development (non-farm) into the watershed, mean of 2002, 2007,
and 2012 National Nutrient Inventory (kg P/ha)}
\item{p_farm_inputs}{Annual total phosphorus inputs from agricultural activities into the watershed, mean of 2002, 2007,
and 2012 National Nutrient Inventory (kg N/ha)}
\item{BFIWs}{Base flow index (percent of total flow that is base flow) for the lake watershed
from LakeCat (percent)}
\item{Precip8110Ws}{30-year mean annual precipitation for the lake watershed, 1981-2010,
from LakeCat (mm)}
\item{Tmean8110Ws}{30-year mean annual air temperature for the lake watershed, 1981-2010,
from LakeCat (per mil)}
\item{fst_ws}{NLCD forested land cover including coniferous, deciduous, and mixed forest.
Data are paired as follows: NLA 2007 - NLCD 2006, NLA 2012 - NLCD 2011,
NLA 2017 - NLCD 2016 (percent)}
\item{MAXDEPTH}{Estimated lake maximum depth from lakemorpho package, NHD+, and LAGOS-US (m)}
\item{lakemorpho_fetch}{Estimated lake fetch from lakemorpho package (m)}
\item{AG_ECO3}{The aggregated 3 ecoregional location of the lake (character)}
\item{lake_area}{Lake surface area in square meters}
\item{Shape}{The centroid of the lake (point)}
\item{pred_cyano_fit}{Predicted cyanobacteria cell abundance (cells/mL) fitted values from 95\% predictiton interval, log transformed: log10(cyano+1000)}
\item{pred_cyano_lwr}{Predicted cyanobacteria cell abundance (cells/mL) lower values from 95\% predictiton interval, log transformed: log10(cyano+1000)}
\item{pred_cyano_upr}{Predicted cyanobacteria cell abundance (cells/mL) upper values from 95\% predictiton interval, log transformed: log10(cyano+1000)}
\item{cyano_transform_fit}{Predicted cyanobacteria cell abundance (cells/mL), fitted values from 95\% predictiton interval}
\item{cyano_transform_lwr}{Predicted cyanobacteria cell abundance (cells/mL), lower values from 95\% predictiton interval}
\item{cyano_transform_upr}{Predicted cyanobacteria cell abundance (cells/mL), upper values from 95\% predictiton interval}
\item{pred_micx_fit}{Predicted risk of microcystin concentration detection above 0.1 ug/L in percent, fitted values from 95\% predictiton interval}
\item{pred_micx_lwr}{Predicted risk of microcystin concentration detection above 0.1 ug/L in percent, lower values from 95\% predictiton interval}
\item{pred_micx_upr}{Predicted risk of microcystin concentration detection above 0.1 ug/L in percent, upper values from 95\% predictiton interval}
\item{micx_transform_fit}{Predicted risk of microcystin concentration detection above 0.1 ug/L, fitter values from 95\% predictiton interval, transformed by 10^(micx) - 1000}
\item{micx_transform_lwr}{Predicted risk of microcystin concentration detection above 0.1 ug/L, lower values from 95\% predictiton interval, transformed by 10^(micx) - 1000}
\item{micx_transform_upr}{Predicted risk of microcystin concentration detection above 0.1 ug/L, upper values from 95\% predictiton interval, transformed by 10^(micx) - 1000}
}
}
\source{
{HABsDrivers} Package for estimating lake HABs risk.
}
\usage{
PredData
}
\description{
A dataset containing data from the National Lakes Assessment as well as
watershed, climate, and lake characteristics. The variables are as follows:
}
\keyword{datasets}
