% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{PredData}
\alias{PredData}
\title{Compiled data for the lake HABs risk modeling project}
\format{
A data frame with 124529 rows and 12 variables:
\describe{
\item{COMID}{A site ID assigned to a lake (character)}
\item{DSGN_CYCLE}{The year of the NLA survey cycle (2007, 2012, 2017)}
\item{n_dev_inputs}{Annual total nitrogen inputs from development (non-farm) into the watershed, mean of 2002, 2007,
and 2012 National Nutrient Inventory (kg N/ha)}
\item{n_farm_inputs}{Annual total nitrogen inputs from agricultural activities into the watershed, mean of 2002, 2007,
and 2012 National Nutrient Inventory (kg N/ha)}
\item{p_dev_inpts}{Annual total phosphorus inputs from development (non-farm) into the watershed, mean of 2002, 2007,
and 2012 National Nutrient Inventory (kg P/ha)}
\item{p_farm_inputs}{Annual total phosphorus inputs from agricultural activities into the watershed, mean of 2002, 2007,
and 2012 National Nutrient Inventory (kg N/ha)}
\item{BFIWs}{Base flow index (percent of total flow that is base flow) for the lake watershed
from LakeCat (percent)}
\item{Precip8110Ws}{30-year mean annual precipitation for the lake watershed, 1981-2010,
from LakeCat (mm)}
\item{Tmean8110Ws}{30-year mean annual air temperature for the lake watershed, 1981-2010,
from LakeCat (per mil)}
\item{fst_ws}{NLCD forested land cover including coniferous, deciduous, and mixed forest.
Data are paired as follows: NLA 2007 - NLCD 2006, NLA 2012 - NLCD 2011,
NLA 2017 - NLCD 2016 (percent)}
\item{MAXDEPTH}{Estimated lake maximum depth from lakemorpho package, NHD+, and LAGOS-US (m)}
\item{lakemorpho_fetch}{Estimated lake fetch from lakemorpho package (m)}
\item{ag_eco3}{The aggregated 3 ecoregional location of the lake (character)}
\item{geometry}{The centroid of the lake (point)}
}
}
\source{
{HABsDrivers} Package for estimating lake HABs risk.
}
\usage{
PredData
}
\description{
A dataset containing data from the National Lakes Assessment as well as
watershed, climate, and lake characteristics. The variables are as follows:
}
\keyword{datasets}
